dot is and cross ts
