+++ content_harvester/derivatives.py
@@ -45,7 +45,10 @@ def make_thumbnail(source_file_path, mimetype):
     if mimetype == 'video/mp4':
         thumbnail = video_to_thumb(source_file_path)
 
-    return thumbnail
+    return {
+        'thumbnail_filepath': thumbnail,
+        'mimetype': 'image/jpeg'
+    }
 
 
 @subprocess_exception_handler
@@ -169,6 +172,7 @@ def tiff_to_srgb_libtiff(input_path):
     return output_path
 
 
+# only used internally to this module
 @subprocess_exception_handler
 def tiff_to_jp2(tiff_path):
     ''' convert a tiff to jp2 using image magick.'''
@@ -196,10 +200,19 @@ def make_jp2(source_file_path, mimetype):
         print(e)
         return source_file_path
 
+    # TODO @bhui: do we still need to prep this file since we are no longer
+    # using kdu_compress? I don't think we do...
     prepped_file_path = None
     if mimetype in ['image/jpeg', 'image/gif', 'image/png', 'image/tiff']:
         converted_file_path = tiff_conversion(source_file_path)
         prepped_file_path = tiff_to_srgb_libtiff(converted_file_path)
+
+        # converted_file_path is an intermediary exclusively used for
+        # prepping the file for jp2 conversion. We can remove it now.
+        # If tiff conversion or tiff_to_srgb_libtiff were no-op, don't delete
+        if (converted_file_path != source_file_path and
+            converted_file_path != prepped_file_path):
+            os.remove(converted_file_path)
 
     if not prepped_file_path:
         print(
@@ -210,4 +223,14 @@ def make_jp2(source_file_path, mimetype):
 
     jp2_filepath = tiff_to_jp2(prepped_file_path)
 
-    return jp2_filepath
+    # prepped_file_path is an intermediary exclusively used for
+    # jp2 conversion. We can remove it now. If tiff to jp2 was a no-op,
+    # don't delete.
+    if (prepped_file_path != source_file_path and
+        jp2_filepath != prepped_file_path):
+        os.remove(prepped_file_path)
+
+    return {
+        'media_filepath': jp2_filepath,
+        'mimetype': 'image/jp2'
+    }
