+++ content_harvester/docker-compose.yml
@@ -9,9 +9,5 @@ services:
     volumes: 
       - ../rikolti_bucket:/rikolti_bucket
       - ./:/content_harvester
-    environment:
-      CONTENT_FETCHER_LOCAL_RUN: "True"
-      CONTENT_FETCHER_DATA_SRC: local
-      CONTENT_FETCHER_DATA_DEST: local
-      NUXEO_BASIC_USER: readonly
-      NUXEO_BASIC_PASS: scat97@gunny
+    env_file:
+      - ../env.local
